Fix up base_formatter
