+++ src/base_formatter/mod.rs
@@ -1103,114 +1103,6 @@ where
     Ok(Formatted::new(document, state.into_context()))
 }
 
-/// Entry point for formatting a [SyntaxNode] for a specific language.
-pub trait FormatLanguage {
-    type SyntaxLanguage: Language;
-
-    /// The type of the formatting context
-    type Context: CstFormatContext<Language = Self::SyntaxLanguage>;
-
-    /// The rule type that can format a [SyntaxNode] of this language
-    type FormatRule: FormatRule<SyntaxNode<Self::SyntaxLanguage>, Context = Self::Context> + Default;
-
-    /// Performs an optional pre-processing of the tree. This can be useful to remove nodes
-    /// that otherwise complicate formatting.
-    ///
-    /// Return [None] if the tree shouldn't be processed. Return [Some] with the transformed
-    /// tree and the source map otherwise.
-    fn transform(
-        &self,
-        _root: &SyntaxNode<Self::SyntaxLanguage>,
-    ) -> Option<SyntaxNode<Self::SyntaxLanguage>> {
-        None
-    }
-
-    /// This is used to select appropriate "root nodes" for the
-    /// range formatting process: for instance in JavaScript the function returns
-    /// true for statement and declaration nodes, to ensure the entire statement
-    /// gets formatted instead of the smallest sub-expression that fits the range
-    fn is_range_formatting_node(&self, _node: &SyntaxNode<Self::SyntaxLanguage>) -> bool {
-        true
-    }
-
-    /// Returns the formatting options
-    fn options(&self) -> &<Self::Context as FormatContext>::Options;
-
-    /// Creates the [FormatContext] with the given `source map` and `comments`
-    fn create_context(self, root: &SyntaxNode<Self::SyntaxLanguage>) -> Self::Context;
-}
-
-/// Formats a syntax node file based on its features.
-///
-/// It returns a [Formatted] result, which the user can use to override a file.
-pub fn format_node<L: FormatLanguage>(
-    root: &SyntaxNode<L::SyntaxLanguage>,
-    language: L,
-) -> FormatResult<Formatted<L::Context>> {
-    let (root, source_map) = match language.transform(&root.clone()) {
-        Some((transformed, source_map)) => {
-            // we don't need to insert the node back if it has the same offset
-            if &transformed == root {
-                (transformed, Some(source_map))
-            } else {
-                match root
-                    .ancestors()
-                    // ancestors() always returns self as the first element of the iterator.
-                    .skip(1)
-                    .last()
-                {
-                    // current root node is the topmost node we don't need to insert the transformed node back
-                    None => (transformed, Some(source_map)),
-                    Some(top_root) => {
-                        // we have to return transformed node back into subtree
-                        let transformed_range = transformed.text_range_with_trivia();
-                        let root_range = root.text_range_with_trivia();
-
-                        let transformed_root = top_root
-                            .replace_child(root.clone().into(), transformed.into())
-                            // SAFETY: Calling `unwrap` is safe because we know that `root` is part of the `top_root` subtree.
-                            .unwrap();
-                        let transformed = transformed_root.covering_element(TextRange::new(
-                            root_range.start(),
-                            root_range.start() + transformed_range.len(),
-                        ));
-
-                        let node = match transformed {
-                            NodeOrToken::Node(node) => node,
-                            NodeOrToken::Token(token) => {
-                                // if we get a token we need to get the parent node
-                                token.parent().unwrap_or(transformed_root)
-                            }
-                        };
-
-                        (node, Some(source_map))
-                    }
-                }
-            }
-        }
-        None => (root.clone(), None),
-    };
-
-    let context = language.create_context(&root);
-    let format_node = FormatRefWithRule::new(&root, L::FormatRule::default());
-
-    let mut state = FormatState::new(context);
-    let mut buffer = VecBuffer::new(&mut state);
-
-    crate::write!(buffer, [format_node])?;
-
-    let mut document = Document::from(buffer.into_vec());
-    document.propagate_expand();
-
-    let context = state.into_context();
-    // let comments = context.comments();
-
-    // comments.assert_checked_all_suppressions(&root);
-    // comments.assert_formatted_all_comments();
-
-    Ok(Formatted::new(document, context))
-}
-
 /// This structure stores the state that is relevant for the formatting of the whole document.
 ///
 /// This structure is different from [crate::base_formatter::Formatter] in that the formatting infrastructure
